+++ src/App.js
@@ -1,12 +1,10 @@
-import TabNavigation from "./Components/TabNavigation/TabNavigation";
-import ContentsSection from "./Components/ContentsSection/ContentsSection";
+import Contents from "./Components/Contents";
 
 export default function App() {
   return (
     <>
       <h1>🤖 프로그래머스 JS 정답 통과기</h1>
-      <TabNavigation />
-      <ContentsSection />
+      <Contents />
     </>
   );
 }
