--- src/Components/ContentsSection/CommonComponents/SubmitButton.js
@@ -1,7 +1,0 @@
-export default function SubmitButton() {
-  return (
-    <div className="SubmitButton">
-      <button id="submitBtn">제출</button>
-    </div>
-  );
-}
