--- src/Components/ContentsSection/ErrorReport/Components/ErrorDetailInput.js
@@ -1,8 +1,0 @@
-export default function ErrorDetailInput() {
-  return (
-    <div className="ErrorDetailInput">
-      <span>내용: </span>
-      <textarea rows="20" cols="100"></textarea>
-    </div>
-  );
-}
