--- src/Components/ContentsSection/SolutionReport/Components/GitHubLogin.js
@@ -1,8 +1,0 @@
-export default function GitHubLogin() {
-  return (
-    <div className="gitHubLogin">
-      <span>기여자 등록: </span>
-      <button id="gitHubLoginBtn">GitHub 로그인</button>
-    </div>
-  );
-}
