--- src/Components/ContentsSection/SolutionReport/Components/SolutionCodeInput.js
@@ -1,8 +1,0 @@
-export default function SolutionCodeInput() {
-  return (
-    <div className="SolutionCodeInput">
-      <span>정답 코드: </span>
-      <textarea rows="20" cols="100"></textarea>
-    </div>
-  );
-}
